committed

+++ app/Http/Controllers/API/UserController.php
@@ -19,7 +19,9 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::latest()->paginate(10);
+        if(\Gate::allows('isAdmin') || \Gate::allows('isAuthor')){
+            return User::latest()->paginate(5);
+        }
     }
 
     /**
@@ -125,8 +127,21 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('isAdmin');
         $user = User::findOrFail($id);
         $user->delete();
         return ['message' => 'User Deleted'];
+    }
+
+    public function search(){
+        if($search = \Request::get('q')){
+            $users = User::where(function($query) use ($search){
+                $query->where('name', 'LIKE', "%$search%")->orWhere('email', 'LIKE', "%$search%")->orWhere('type', 'LIKE', "%$search%");
+            })->paginate(10);
+        }else{
+            $users = User::latest()->paginate(5);
+        }
+
+        return $users;
     }
 }
